+++ tests/testthat/test-paired_test_continuous.R
@@ -1,0 +1,29 @@
+context("test-paired_test_continuous")
+
+iris <- dplyr::group_by(iris, Species) %>%
+  dplyr::mutate(id = dplyr::row_number()) %>%
+  dplyr::ungroup() %>%
+  dplyr::arrange(id)
+
+which_selected <- iris$Species != "setosa"
+two_obs    <- iris$Sepal.Length[which_selected]
+two_groups <- iris$Species[which_selected] %>% droplevels()
+
+obs <- iris$Sepal.Length
+many_groups <- iris$Species
+
+test_that("output class is correct", {
+  expect_is(paired_test_continuous(two_groups, two_obs), "list")
+  expect_is(paired_test_continuous(many_groups, obs), "list")
+})
+
+test_that("output structure is correct", {
+  expect(all(
+    c("P", "stat", "df", "testname", "statname", "namefun") %in%
+    names(paired_test_continuous(two_groups, two_obs))
+  ))
+  expect(all(
+    c("P", "stat", "df", "testname", "statname", "namefun") %in%
+    names(paired_test_continuous(many_groups, obs))
+  ))
+})